Refactor SplitText
Refactor LCModelComponent status message formatting for better readability
